refactor verify pow target
