+++ core/consensus/pow/src/job.rs
@@ -41,7 +41,7 @@ use {
 	inherents::InherentDataProviders,
 	runtime_primitives::{
 		generic::BlockId,
-		traits::{Block, ProvideRuntimeApi, Zero, One, As, DigestItemFor, NumberFor, SimpleArithmetic, DigestFor, Digest, Header},
+		traits::{Block, ProvideRuntimeApi, As, DigestItemFor, NumberFor, SimpleArithmetic, DigestFor, Digest, Header},
 	},
 };
 use {
@@ -56,7 +56,7 @@ use log::info;
 use {
 	pow_primitives::{YeePOWApi, PowTarget},
 };
-use crate::pow::{check_proof, gen_extrinsic_proof};
+use crate::pow::{check_proof, gen_extrinsic_proof, calc_pow_target};
 use yee_sharding::{ShardingDigestItem, ScaleOutPhaseDigestItem};
 use crate::verifier::check_shard_info;
 use primitives::H256;
@@ -253,70 +253,4 @@ impl<B, C, E, AccountId, AuthorityId, I> JobManager for DefaultJobManager<B, C, 
 fn timestamp_now() -> Result<u64, consensus_common::Error> {
 	Ok(SystemTime::now().duration_since(UNIX_EPOCH)
 		.map_err(to_common_error)?.as_millis() as u64)
-}
-
-fn calc_pow_target<B, C, AuthorityId>(
-	client: Arc<C>, header: &<B as Block>::Header, timestamp: u64,
-) -> Result<PowTarget, consensus_common::Error> where
-	B: Block,
-	NumberFor<B>: SimpleArithmetic,
-	DigestFor<B>: Digest,
-	DigestItemFor<B>: super::CompatibleDigestItem<B, AuthorityId>,
-	C: HeaderBackend<B> + ProvideRuntimeApi,
-	<C as ProvideRuntimeApi>::Api: YeePOWApi<B>,
-	AuthorityId: Encode + Decode + Clone,
-{
-	let next_num = *header.number();
-	let curr_block_id = BlockId::hash(*header.parent_hash());
-	let api = client.runtime_api();
-	let genesis_pow_target = api.genesis_pow_target(&curr_block_id)
-		.map_err(to_common_error)?;
-	let adj = api.pow_target_adj(&curr_block_id)
-		.map_err(to_common_error)?;
-	let curr_header = client.header(curr_block_id)
-		.expect("parent block must exist for sealer; qed")
-		.expect("parent block must exist for sealer; qed");
-	let one = <NumberFor<B> as As<u64>>::sa(1u64);
-	// not on adjustment, reuse parent pow target
-	if next_num == one {
-		return Ok(genesis_pow_target)
-	} else if (next_num - one) % adj != Zero::zero() {
-		let curr_pow_target = curr_header.digest().logs().iter().rev()
-			.filter_map(CompatibleDigestItem::as_pow_seal).next()
-			.and_then(|seal| Some(seal.pow_target))
-			.unwrap_or(genesis_pow_target);
-		return Ok(curr_pow_target);
-	}
-
-	let curr_seal = curr_header.digest().logs().iter().rev()
-		.filter_map(CompatibleDigestItem::as_pow_seal).next()
-		.expect("Seal must exist when adjustment comes; qed");
-	let curr_pow_target = curr_seal.pow_target;
-
-	let (block_gap, last_time) = {
-		let id = BlockId::<B>::number(next_num - adj);
-		let ancestor_header = client.header(id)
-			.expect("parent block must exist for sealer; qed")
-			.expect("parent block must exist for sealer; qed");
-		let ancestor_seal = ancestor_header.digest().logs().iter().rev()
-			.filter_map(CompatibleDigestItem::as_pow_seal).next();
-		match ancestor_seal {
-			Some(seal) => {
-				(adj.as_(), seal.timestamp)
-			}
-			None => {
-				panic!("can't get PowSeal in pre-block's header")
-			}
-		}
-	};
-
-	let target_block_time = api.target_block_time(&curr_block_id)
-		.map_err(to_common_error)?;
-	let time_gap = timestamp - last_time;
-	let expected_gap = target_block_time * 1000 * block_gap;
-	let new_pow_target = (curr_pow_target / expected_gap) * time_gap;
-	info!("pow target adjustment: gap: {}, time: {}", block_gap, time_gap);
-	info!("old pow target: {:#x}, new pow target: {:#x}",curr_pow_target, new_pow_target);
-
-	Ok(new_pow_target)
 }
